+++ HISTORY.md
@@ -109,3 +109,14 @@ https://start.spring.io/
 @Bean이 붙은 메서드마다 이미 스프링 빈이 존재하면 존재하는 빈을 반환, 스프링 빈이 없으면 생성해서 스프링 빈으로 등록하고 반환하는 코드가 동적으로 만든다.   
 따라서, @Configuration없이 @Bean만 사용할 경우 스프링 빈에 등록은 되지만 싱글톤을 보장해주지는 않는다.
 ---
+###2024-01-22
+####컴포넌트 스캔
+@ComponentScan 은 @Component 가 붙은 모든 클래스를 스프링 빈으로 등록한다.  
+이때 스프링 빈의 기본 이름은 클래스명을 사용하되 맨 앞글자만 소문자를 사용한다.  
+* 빈 이름 기본 전략: MemberServiceImpl 클래스 memberServiceImpl  
+* 빈 이름 직접 지정: 만약 스프링 빈의 이름을 직접 지정하고 싶으면 @Component("memberService2") 이런식으로 이름을 부여하면 된다
+####@Autowired 의존관계 자동 주입
+생성자에 @Autowired 를 지정하면, 스프링 컨테이너가 자동으로 해당 스프링 빈을 찾아서 주입한다.  
+이때 기본 조회 전략은 타입이 같은 빈을 찾아서 주입한다.  
+getBean(MemberRepository.class) 와 동일하다고 이해하면 된다
+---